Rename LanguageSupport module to Languages
