--- lib/pragmatic_segmenter/language_support.rb
@@ -1,33 +1,0 @@
-require 'pragmatic_segmenter/languages'
-
-module PragmaticSegmenter
-  module LanguageSupport
-    LANGUAGE_CODES = {
-      'en' => 'English',
-      'de' => 'Deutsch',
-      'es' => 'Spanish',
-      'fr' => 'French',
-      'it' => 'Italian',
-      'ja' => 'Japanese',
-      'el' => 'Greek',
-      'ru' => 'Russian',
-      'ar' => 'Arabic',
-      'am' => 'Amharic',
-      'hi' => 'Hindi',
-      'hy' => 'Armenian',
-      'fa' => 'Persian',
-      'my' => 'Burmese',
-      'ur' => 'Urdu',
-      'nl' => 'Dutch',
-      'pl' => 'Polish',
-    }
-
-    def process_class
-      Object.const_get("PragmaticSegmenter::Languages::#{LANGUAGE_CODES[language] || 'Common'}::Process")
-    end
-
-    def cleaner_class
-      Object.const_get("PragmaticSegmenter::Languages::#{LANGUAGE_CODES[language] || 'Common'}::Cleaner")
-    end
-  end
-end
